requirements.txt
fastapi[standard]
uvicorn
strawberry-graphql
sqlalchemy
sqlalchemy-orm
databases
dataclasses
